requirements.txt
Django==1.11.1

djangorestframework==3.6.3
